requirements.txt
pyyaml
discord.py
requests
Django
djangorestframework
markdown
django-filter
djangorestframework-api-key==2.
